* Fix bug with drawing labels: after font->glWorldPrint   the InMapDraw texture wasn't reset. * Fix gcc warnings.
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@1426 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Sim/Projectiles/ExplosionGenerator.cpp
@@ -302,7 +302,7 @@ void CCustomExplosionGenerator::ExecuteExplosionCode (const char *code, float da
 		case OP_STOREI:{
 			Uint16 offset = *(Uint16*)code;
 			code += 2;
-			*(int*)(instance + offset) = val;
+			*(int*)(instance + offset) = (int) val;
 			val = 0.0f;
 			break;}
 		case OP_STOREF:{
@@ -314,7 +314,7 @@ void CCustomExplosionGenerator::ExecuteExplosionCode (const char *code, float da
 		case OP_STOREC:{
 			Uint16 offset = *(Uint16*)code;
 			code += 2;
-			*(unsigned char*)(instance + offset) = val;
+			*(unsigned char*)(instance + offset) = (int) val;
 			val = 0.0f;
 			break;}
 		case OP_ADD:
